requirements.txt
streamlit==1.32.0
websocket-client==1.7.0
requests==2.31.0 